requirements.txt
flask
python-dotenv
anthropic
googlemaps
requests
beautifulsoup4
playwright
